requirements.txt
fastapi
uvicorn
pdfplumber
groq
python-multipart
python-dotenv
